Rework linked list class
Object methods are now on the prototype. This way less memory is used
because there isn't functions in every instance.

Also working with private class fields and properties so the head of
the linked list isn't available on the instance publicly.
(It's only accessible on the instance only in the class definition)